number of players decreases after one player is gone and game instructions in progress
